+++ .vscode/tasks.json
@@ -1,0 +1,15 @@
+{
+  // See https://go.microsoft.com/fwlink/?LinkId=733558
+  // for the documentation about the tasks.json format
+  "version": "2.0.0",
+  "tasks": [
+    {
+      "label": "go build",
+      "type": "shell",
+      "command": "go build",
+      "options": {
+        "cwd": "${workspaceFolder}/cmd/s3backup"
+      }
+    }
+  ]
+}
