+++ go.mod
@@ -1,12 +1,11 @@
-module s3-backup
+module github.com/ntsaini/s3-backup
 
-go 1.19
-
-require gopkg.in/yaml.v2 v2.2.8
+go 1.20
 
 require (
-	github.com/aws/aws-sdk-go v1.44.189
+	github.com/aws/aws-sdk-go v1.44.199
 	github.com/bmatcuk/doublestar v1.3.4
+	gopkg.in/yaml.v2 v2.4.0
 )
 
 require github.com/jmespath/go-jmespath v0.4.0 // indirect
